pyproject.toml
[project]
name = "newgcal"
version = "0.1.0"
description = "Multi-agent GCP Cloud Calculator using Google ADK"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.31.1",
    "google-cloud-billing>=1.19.0",
    "google-auth>=2.37.0",
    "requests>=2.32.3",
    "python-dotenv>=1.2.2",
]
